pyproject.toml
[build-system]
requires = [
    "setuptools >= 59.0",
    "setuptools_scm",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "clp_ffi_py"
version = "0.0.12"
license = { text = "Apache License 2.0" }
authors = [
    { name="zhihao lin", email="lin.zhihao@yscope.com" },
]
description = "Python interface to the CLP Core Features through CLP's FFI"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "Deprecated >= 1.2.14",
    "python-dateutil >= 2.7.0",
    "typing-extensions >= 4.1.1",
    "zstandard >= 0.18.0",
]
classifiers = [
"License :: OSI Approved :: Apache Software License",
]

[project.urls]
"Homepage" = "https://github.com/y-scope/clp-ffi-py"
"Bug Tracker" = "https://github.com/y-scope/clp-ffi-py/issues"

[tool.black]
line-length = 100
target-version = ["py311"]
color = true
preview = true

[tool.cibuildwheel]
# NOTE: The wheels to be built (and skipped) must be kept in-sync with the GitHub workflow. The GH
# workflow doesn't use pyproject.toml to determine which wheels to build, but instead uses
# environment variables so that it can build the wheels in parallel jobs.

# Use `musllinux_1_2` to support C++20 compilation
musllinux-aarch64-image = "musllinux_1_2"
musllinux-i686-image = "musllinux_1_2"
musllinux-ppc64le-image = "musllinux_1_2"
musllinux-s390x-image = "musllinux_1_2"
musllinux-x86_64-image = "musllinux_1_2"

skip = "pp*"

test-command = [
  "python -m unittest discover --failfast --verbose --start-directory={package}/tests"
]
test-requires = [
    "smart_open==6.4.0"
]
test-skip = [
    "cp39-*",
    "cp310-*",
    "*-*linux_i686",
]

[tool.cibuildwheel.linux]
archs = ["auto", "aarch64"]

[tool.cibuildwheel.macos]
archs = ["x86_64", "universal2", "arm64"]

[tool.cibuildwheel.windows]
archs = ["AMD64"]

[tool.docformatter]
make-summary-multi-line = true
pre-summary-newline = true
recursive = true
wrap-summaries = 100
wrap-descriptions = 100

[tool.mypy]
explicit_package_bases = true
mypy_path = ["tests"]
strict = true
pretty = true

[tool.ruff]
exclude = [
    "clp_ffi_py/ir/__init__.py"
]
line-length = 100

[tool.ruff.lint]
select = ["E", "I", "F"]
isort.order-by-type = false
